+++ test/cookbooks/test/recipes/mod_ssl.rb
@@ -8,7 +8,7 @@ apache2_install 'default'
 service 'apache2' do
   service_name lazy { apache_platform_service_name }
   supports restart: true, status: true, reload: true
-  action [:start, :enable]
+  action :nothing
 end
 
 apache2_module 'deflate'