+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file is used to list changes made in each version of the apache2 cookbook.
 
+## Unreleased
+
 ## 8.10.0 - *2021-04-09*
 
 - Fix `apache2_mod_auth_cas` resource for all supported platforms