+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file is used to list changes made in each version of the apache2 cookbook.
 ## Unreleased
 
 - Add missing unified_mode from mod_php and mod_wscgi
+- Fix service resource restarting the service every run
 
 ## 8.10.0 - *2021-04-09*