+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This file is used to list changes made in each version of the apache2 cookbook.
 
-## Unreleased
+## 8.11.0 - *2021-05-06*
 
 - Add missing unified_mode from mod_php and mod_wscgi
 - Fix service resource restarting the service every run